+++ src/View/NugetMonitorViewModel.cs
@@ -4,55 +4,12 @@ using System.Windows.Input;
 using Community.VisualStudio.Toolkit;
 using Microsoft.Build.Construction;
 using Microsoft.Build.Evaluation;
-using NuGet.Packaging.Core;
-using NuGet.Versioning;
-using NuGetMonitor.Models;
+using Microsoft.VisualStudio.Shell;
 using NuGetMonitor.Services;
 using TomsToolbox.Essentials;
 using TomsToolbox.Wpf;
 
 namespace NuGetMonitor.View;
-
-internal partial class PackageViewModel : INotifyPropertyChanged
-{
-    private readonly NugetMonitorViewModel _owner;
-
-    public PackageViewModel(IGrouping<PackageIdentity, PackageReferenceEntry> items, NugetMonitorViewModel owner)
-    {
-        _owner = owner;
-        Items = items;
-        Identity = items.Key;
-        ProjectPaths = string.Join(", ", items.Select(item => item.RelativePath));
-    }
-
-    public IGrouping<PackageIdentity, PackageReferenceEntry> Items { get; }
-
-    public PackageIdentity Identity { get; private set; }
-
-    public string ProjectPaths { get; }
-
-    public async Task Load()
-    {
-        Package = await NuGetService.GetPackage(Identity.Id).ConfigureAwait(false);
-        SelectedVersion = Package.Versions.FirstOrDefault(i => !i.IsPrerelease && i >= Identity.Version) ?? Package.Versions.FirstOrDefault();
-        IsSelected = IsUpdateAvailable;
-    }
-
-    public Package? Package { get; private set; }
-
-    public NuGetVersion? SelectedVersion { get; set; }
-
-    public bool IsUpdateAvailable => SelectedVersion > Identity.Version;
-
-    public bool IsSelected { get; set; }
-
-    public ICommand Update => new DelegateCommand(() => { _owner.Update(this); });
-
-    public void ApplyVersion()
-    {
-        Identity = new PackageIdentity(Identity.Id, SelectedVersion);
-    }
-}
 
 internal partial class NugetMonitorViewModel : INotifyPropertyChanged
 {
@@ -61,12 +18,32 @@ internal partial class NugetMonitorViewModel : INotifyPropertyChanged
     private NugetMonitorViewModel()
     {
         VS.Events.SolutionEvents.OnAfterOpenSolution += SolutionEvents_OnAfterOpenSolution;
+        VS.Events.SolutionEvents.OnAfterCloseSolution += SolutionEvents_OnAfterCloseSolution;
         Load();
     }
+
+    public ObservableCollection<PackageViewModel> Packages { get; } = new();
+
+    public ObservableCollection<PackageViewModel> SelectedPackages { get; } = new();
+
+    public bool IsLoading { get; set; } = true;
+
+    public ICommand UpdateSelectedCommand => new DelegateCommand(UpdateSelected);
+
+    public ICommand SoftRefreshCommand => new DelegateCommand(SoftRefresh);
+
+    public ICommand HardRefreshCommand => new DelegateCommand(HardRefresh);
+    
+    public ICommand ShowNuGetPackageManagerCommand => new DelegateCommand(ShowNuGetPackageManager);
 
     private void SolutionEvents_OnAfterOpenSolution(Solution? obj)
     {
         Load();
+    }
+
+    private void SolutionEvents_OnAfterCloseSolution()
+    {
+        Packages.Clear();
     }
 
     private async void Load()
@@ -98,27 +75,20 @@ internal partial class NugetMonitorViewModel : INotifyPropertyChanged
         }
     }
 
-    public ObservableCollection<PackageViewModel> Packages { get; } = new();
-
-    public ObservableCollection<PackageViewModel> SelectedPackages { get; } = new();
-
-    public bool IsLoading { get; set; } = true;
-
-    public ICommand UpdateSelectedCommand => new DelegateCommand(UpdateSelected);
-
-    public ICommand SoftRefreshCommand => new DelegateCommand(SoftRefresh);
-
-    public ICommand HardRefreshCommand => new DelegateCommand(HardRefresh);
+    private static void ShowNuGetPackageManager()
+    {
+        VS.Commands.ExecuteAsync("Tools.ManageNuGetPackagesForSolution").FireAndForget();
+    }
 
     private void HardRefresh()
     {
         NuGetService.ClearCache();
         Load();
     }
-    
+
     private void SoftRefresh()
     {
-        Load();     
+        Load();
     }
 
     public void Update(PackageViewModel packageViewModel)
